refactor: update backend logic and adjust README
